refactor: favorite item 도메인과 item 도메인 기능 분리
