+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/api/dto/response/FolderGetByCursorResponse.java
@@ -1,8 +1,8 @@
-package com.programmers.lime.domains.item.api.dto.response;
+package com.programmers.lime.domains.favoriteItem.api.dto.response;
 
 import java.util.List;
 
-import com.programmers.lime.domains.item.application.dto.FolderGetByCursorServiceResponse;
+import com.programmers.lime.domains.favoriteItem.application.dto.FolderGetByCursorServiceResponse;
 import com.programmers.lime.domains.item.model.MemberItemFolderCursorSummary;
 
 public record FolderGetByCursorResponse(
