+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/api/dto/response/MemberItemCreateResponse.java
@@ -1,8 +1,8 @@
-package com.programmers.lime.domains.item.api.dto.response;
+package com.programmers.lime.domains.favoriteItem.api.dto.response;
 
 import java.util.List;
 
-import com.programmers.lime.domains.item.application.dto.MemberItemCreateServiceResponse;
+import com.programmers.lime.domains.favoriteItem.application.dto.MemberItemCreateServiceResponse;
 
 public record MemberItemCreateResponse(List<Long> itemIds) {
 	public static MemberItemCreateResponse from(final MemberItemCreateServiceResponse response) {
