+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/api/dto/response/MemberItemFolderGetByCursorResponse.java
@@ -1,9 +1,9 @@
-package com.programmers.lime.domains.item.api.dto.response;
+package com.programmers.lime.domains.favoriteItem.api.dto.response;
 
 import java.util.List;
 
 import com.programmers.lime.common.cursor.CursorSummary;
-import com.programmers.lime.domains.item.application.dto.MemberItemFolderGetServiceResponse;
+import com.programmers.lime.domains.favoriteItem.application.dto.MemberItemFolderGetServiceResponse;
 import com.programmers.lime.domains.item.model.MemberItemFolderCursorSummary;
 
 public record MemberItemFolderGetByCursorResponse (
