+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/application/dto/FolderGetByCursorServiceResponse.java
@@ -1,4 +1,4 @@
-package com.programmers.lime.domains.item.application.dto;
+package com.programmers.lime.domains.favoriteItem.application.dto;
 
 import com.programmers.lime.common.cursor.CursorSummary;
 import com.programmers.lime.domains.item.model.MemberItemFolderCursorSummary;
