+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/application/dto/MemberItemCreateServiceResponse.java
@@ -1,4 +1,4 @@
-package com.programmers.lime.domains.item.application.dto;
+package com.programmers.lime.domains.favoriteItem.application.dto;
 
 import java.util.List;
