+++ lime-api/src/main/java/com/programmers/lime/domains/favoriteItem/application/dto/MemberItemFolderGetServiceResponse.java
@@ -1,8 +1,7 @@
-package com.programmers.lime.domains.item.application.dto;
+package com.programmers.lime.domains.favoriteItem.application.dto;
 
 import com.programmers.lime.common.cursor.CursorSummary;
 import com.programmers.lime.domains.item.model.MemberItemFolderCursorSummary;
-import com.programmers.lime.domains.item.model.MemberItemFolderSummary;
 
 public record MemberItemFolderGetServiceResponse(
